requirements.txt
# These requirements are for running tests, the library itself only needs a
# subset, see setup.py for that list
click==8.1.3
click-aliases==1.0.1
click-option-group==0.5.5
colorama==0.4.6
coverage==7.0.1
gitpython==3.1.30
pudb==2022.1.3
pyjwt==2.6.0
pyotp==2.8.0
python-gnupg>=0.3.5,<0.5.0  # Note the updated version range
requests==2.28.1
tabulate==0.9.0
tqdm==4.64.1
